requirements.txt
pillow~=10.0.1
selenium~=4.16.0
requests~=2.31.0
pymysql~=1.1.0